refactor: eslint package structure
