+++ packages/eslint-config/src/base/astro.ts
