+++ packages/eslint-config/src/base/js.ts
