+++ packages/eslint-config/src/base/nextjs.ts
@@ -1,0 +1,11 @@
+import type { Linter } from "eslint"
+
+import { compat } from "../lib/compat"
+import { reactConfig } from "./react"
+
+const nextJsConfig: Linter.FlatConfig[] = [
+  ...reactConfig,
+  ...compat.extends("next/core-web-vitals"),
+]
+
+export { nextJsConfig }
