+++ packages/eslint-config/src/base/react.ts
