+++ packages/eslint-config/src/base/stylistic.ts
