+++ packages/eslint-config/src/base/typescript.ts
