+++ packages/eslint-config/src/index.ts
@@ -1,4 +1,1 @@
-export { default as astro } from "./astro"
-export { default as js } from "./js"
-export { default as react } from "./react"
-export { default as ts } from "./ts"
+export { compat } from "./lib/compat"
