--- packages/eslint-config/src/js.ts
@@ -1,6 +1,0 @@
-import { jsConfig } from "./bases/js"
-import { stylisticConfig } from "./bases/stylistic"
-
-const js = [...jsConfig, ...stylisticConfig]
-
-export default js
