+++ packages/eslint-config/src/presets/index.ts
@@ -1,0 +1,13 @@
+import astro from "./astro"
+import js from "./js"
+import react from "./react"
+import ts from "./ts"
+
+const presets = {
+  astro,
+  js,
+  react,
+  ts,
+}
+
+export default presets
