+++ packages/eslint-config/src/presets/js.ts
@@ -1,0 +1,6 @@
+import { jsConfig } from "../base/js"
+import { stylisticConfig } from "../base/stylistic"
+
+const js = [...jsConfig, ...stylisticConfig]
+
+export default js
