+++ packages/eslint-config/src/presets/nextjs.ts
@@ -2,12 +2,12 @@ import type { TSESLint } from "@typescript-eslint/utils"
 
 import tseslint from "typescript-eslint"
 
-import { reactConfig } from "./bases/react"
+import { nextJsConfig } from "../base/nextjs"
 import ts from "./ts"
 
 const react: TSESLint.FlatConfig.ConfigArray = tseslint.config(
   ...ts,
-  ...reactConfig,
+  ...nextJsConfig,
 )
 
 export default react
