+++ packages/eslint-config/src/presets/react.ts
@@ -1,0 +1,13 @@
+import type { TSESLint } from "@typescript-eslint/utils"
+
+import tseslint from "typescript-eslint"
+
+import { reactConfig } from "../base/react"
+import ts from "./ts"
+
+const react: TSESLint.FlatConfig.ConfigArray = tseslint.config(
+  ...ts,
+  ...reactConfig,
+)
+
+export default react
