+++ packages/eslint-config/tsup.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from "tsup"
 
 export default defineConfig({
   clean: true, // build前にディレクトリ内を削除するか
-  entry: ["./src/*.ts", "./src/addons/*.ts"], // バンドルするファイルを指定
+  entry: ["./src/*.ts", "./src/presets/*.ts", "./src/addons/*.ts"], // バンドルするファイルを指定
   experimentalDts: true,
   format: ["esm"], // 出力する形式を指定
   minify: process.env.NODE_ENV === "production",
